fix(test-isolation): share fake Firestore transactional helper
